Added project and initial game code
Added several classes, the main ones are listed below:
CMaze - a little class that build a minimal spanning tree and returns
the walls in run length list.. This class should be extended to also
return a trinary(?) tree depicting the graph from any given cell
position - this would be useful for player and AI placement in equal
distance from target as well as quickly determining the AI's path, and
avoiding the need to use collision checks on the mesh level.
CMazeGameEngine - this is the big all knowing class - it holds the
camera, player, (will hold AI) scene-graph, etc.
CMazePlayer - This is a representation of an actor in the game whether
he's controlled by AI or keyboard - In it's current form the graphics
engine is to invasive - This class would be refactored to be independent
from the engine...